# docs/swagger.yaml
basePath: /v1/2024
definitions:
...
  model.AddNewEventParams:
    properties:
      description:
        type: string
      end:
        description: ISO 8601
        type: string
      image:
        type: string
      name:
        type: string
      start:
        description: 'ISO 8601: "2025-04-08T17:30:00Z"'
        type: string
    type: object
...
  model.NewOrUpdateMenuFunctionParams:
    properties:
      active:
        type: boolean
      description:
        type: string
      name:
        type: string
      url:
        type: string
    type: object
...
  model.UpdateEventParams:
    properties:
      active:
        type: boolean
      description:
        type: string
      end:
        type: string
      image:
        type: string
      name:
        type: string
      start:
        type: string
      user_id:
        type: string
    type: object
...
  response.ErrResponse:
    properties:
      code:
        type: integer
      detail: {}
      messgae:
        type: string
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      messgae:
        type: string
    type: object
host: localhost:8082
info:
  contact:
    email: quang0706r@gmail.com
    name: TEAM QUANG
    url: https://github.com/Quangnguyen1505/Ecommerce-Go
  description: This is a server ecommerce.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/Quangnguyen1505/Ecommerce-Go
  title: Demo API Ecommerce
  version: 1.0.0
paths:
...
  /events:
    get:
      consumes:
      - application/json
      description: When admin wants to get all event content
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get all events
      tags:
      - event
    post:
      consumes:
      - application/json
      description: When admin wants to add a new event
      parameters:
      - description: authorization token
        in: header
        name: authorization
        required: true
        type: string
      - description: x-client-id user
        in: header
        name: x-client-id
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.AddNewEventParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Create new event
      tags:
      - event
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: When admin wants to delete an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Delete event
      tags:
      - event
    get:
      consumes:
      - application/json
      description: When admin wants to get an event by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get event by ID
      tags:
      - event
    put:
      consumes:
      - application/json
      description: When admin wants to update an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateEventParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Update event
      tags:
      - event
  /menu-function/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a menu function by its ID
      parameters:
      - description: Menu Function ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Delete menu function
      tags:
      - menu function
  /menu-functions:
    get:
      consumes:
      - application/json
      description: Get all menu functions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get all menu functions
      tags:
      - menu function
    post:
      consumes:
      - application/json
      description: Create a new menu function
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.NewOrUpdateMenuFunctionParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Create new menu function
      tags:
      - menu function
  /menu-functions/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific menu function by its ID
      parameters:
      - description: Menu Function ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get menu function by ID
      tags:
      - menu function
    put:
      consumes:
      - application/json
      description: Update an existing menu function
      parameters:
      - description: Menu Function ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.NewOrUpdateMenuFunctionParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Update menu function
      tags:
      - menu function
  /menu-functions/active:
    get:
      consumes:
      - application/json
      description: Get all menu functions that are currently active
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Get all active menu functions
      tags:
      - menu function
...
swagger: "2.0"